@@ -10,53 +10,53 @@ describe LinkCheckRulesController do
     end
 
     describe 'GET index' do
-      it 'should get all of the collection rules' do
-        LinkCheckRule.should_receive(:all) { [link_check_rule] }
+      it 'returns all of the collection rules' do
+        allow(LinkCheckRule).to receive(:all).and_return([link_check_rule])
         get :index, params: {}, format: 'json'
-        assigns(:link_check_rules).should eq [link_check_rule]
+        expect(assigns(:link_check_rules)).to eq [link_check_rule]
       end
 
-      it 'should do a where if link_check_rule is defined' do
+      it 'performs a where if link_check_rule is defined' do
         params = { link_check_rule: { collection_title: 'link_check_rule' } }
-        LinkCheckRule.should_receive(:where).with(params[:link_check_rule].stringify_keys)
+        expect(LinkCheckRule).to receive(:where).with(params[:link_check_rule].stringify_keys)
         get :index, params: params, format: 'json'
       end
     end
 
     describe 'GET show' do
-      it 'should get the collection rule' do
-        LinkCheckRule.should_receive(:find) { link_check_rule }
+      it 'finds the collection rule' do
+        allow(LinkCheckRule).to receive(:find).and_return(link_check_rule)
         get :show, params: { id: link_check_rule.id }
-        assigns(:link_check_rule).should eq link_check_rule
+        expect(assigns(:link_check_rule)).to eq link_check_rule
       end
     end
 
     describe 'POST create' do
-      it 'should make a new collection rule and assign it' do
+      it 'creates a new collection rule and assign it' do
         params = ActionController::Parameters.new('collection_title' => 'collection_title', 'status_codes' => '203,205', source_id: 'source_id').permit!
-        LinkCheckRule.should_receive(:create!).with(params) { link_check_rule }
+        allow(LinkCheckRule).to receive(:create!).with(params).and_return(link_check_rule)
         post :create, params: { link_check_rule: { collection_title: 'collection_title', status_codes: '203,205', source_id: 'source_id' } }
-        assigns(:link_check_rule) { link_check_rule }
+        expect(assigns(:link_check_rule)).to eq link_check_rule
       end
     end
 
     describe 'PUT update' do
-      it 'should find the link_check_rule' do
-        LinkCheckRule.should_receive(:find) { link_check_rule }
+      it 'finds the link_check_rule' do
+        allow(LinkCheckRule).to receive(:find).and_return(link_check_rule)
         put :update, params: { id: link_check_rule.id, link_check_rule: { collection_title: 'collection_title', status_codes: '203,205' } }
-        assigns(:link_check_rule) { link_check_rule }
+        expect(assigns(:link_check_rule)).to eq link_check_rule
       end
 
       it 'updates all the attributes' do
-        LinkCheckRule.stub(:find) { link_check_rule }
-        link_check_rule.should_receive(:update).with('collection_title' => 'collection_title', 'status_codes' => '203,205')
+        allow(LinkCheckRule).to receive(:find).and_return(link_check_rule)
+        expect(link_check_rule).to receive(:update).with('collection_title' => 'collection_title', 'status_codes' => '203,205')
         put :update, params: { id: link_check_rule.id, link_check_rule: { collection_title: 'collection_title', status_codes: '203,205' } }
       end
     end
 
     describe 'DELETE destroy' do
       it 'finds the collection rule and destroys it' do
-        LinkCheckRule.should_receive(:find) { link_check_rule }
+        expect(LinkCheckRule).to receive(:find).and_return(link_check_rule)
         delete :destroy, params: { id: link_check_rule.id }
       end
     end
