@@ -7,19 +7,19 @@ describe PreviewsController do
 
   describe 'POST create' do
     let(:preview) { double(:preview) }
-    before { Preview.stub(:spawn_preview_worker) { 'abc123' } }
+    before { allow(Preview).to receive(:spawn_preview_worker).and_return('abc123') }
 
     it 'spawns a preview worker' do
-      Preview.should_receive(:spawn_preview_worker).with('id' => '123', 'harvest_job' => { 'parser_code' => 'CODE', 'index' => '1000' })
+      allow(Preview).to receive(:spawn_preview_worker).with('id' => '123', 'harvest_job' => { 'parser_code' => 'CODE', 'index' => '1000' })
       post :create, params: { preview: { id: '123', harvest_job: { parser_code: 'CODE', index: 1000 } } }
     end
   end
 
   describe 'GET show' do
     it 'should find the preview object' do
-      Preview.should_receive(:find).with(preview.id.to_s) { preview }
+      allow(Preview).to receive(:find).with(preview.id.to_s).and_return(preview)
       get :show, params: { id: preview.id }
-      assigns(:preview).should eq preview
+      expect(assigns(:preview)).to eq preview
     end
   end
 end
