@@ -9,42 +9,43 @@ describe AbstractJob do
 
     it 'returns all active harvest jobs' do
       create(:abstract_job, status: 'finished')
-      AbstractJob.search('status' => 'active').should eq [active_job]
+      expect(AbstractJob.search('status' => 'active')).to eq [active_job]
     end
 
     it 'paginates through the records' do
-      AbstractJob.should_receive(:page).with(2) { AbstractJob.unscoped }
+      expect(AbstractJob).to receive(:page).with(2).and_return(AbstractJob.unscoped)
       AbstractJob.search('status' => 'active', 'page' => '2').to_a
     end
 
     it 'returns the recent harvest jobs first' do
       active_job2 = create(:abstract_job, status: 'active', start_time: Time.now + 5.seconds)
-      AbstractJob.search('status' => 'active').first.should eq active_job2
+      expect(AbstractJob.search('status' => 'active').first).to eq active_job2
     end
 
     it 'returns only test harvest jobs of a specific parser' do
       job2 = create(:abstract_job, parser_id: '333', environment: 'test')
-      AbstractJob.search('parser_id' => '333', 'environment' => 'test').should eq [job2]
+      abstract_job_search = AbstractJob.search('parser_id' => '333', 'environment' => 'test')
+      expect(abstract_job_search).to eq [job2]
     end
 
     it 'limits the number of harvest jobs returned' do
       create(:abstract_job, parser_id: '333', environment: 'test', start_time: Time.now + 5.seconds)
-      AbstractJob.search('limit' => '1').to_a.size.should eq 1
+      abstract_job_results = AbstractJob.search('limit' => '1').to_a.size
+      expect(abstract_job_results).to eq 1
     end
 
-    it 'should find all harvest jobs either in staging or production' do
+    it 'finds all harvest jobs either in staging or production' do
       job1 = create(:abstract_job, parser_id: '333', environment: 'staging', start_time: Time.now)
       job2 = create(:abstract_job, parser_id: '334', environment: 'production', start_time: Time.now + 2.seconds)
       jobs = AbstractJob.search('environment' => %w[staging production]).to_a
-      jobs.should include(job2)
-      jobs.should include(job1)
+      expect(jobs).to include job1, job2
     end
   end
 
   describe '.clear_raw_data' do
-    it 'should fetch harvest jobs older than a week' do
-      AbstractJob.should_receive(:disposable) { [job] }
-      job.should_receive(:clear_raw_data)
+    it 'fetches harvest jobs older than a week' do
+      expect(AbstractJob).to receive(:disposable).and_return([job])
+      expect(job).to receive(:clear_raw_data)
       AbstractJob.clear_raw_data
     end
   end
@@ -67,12 +68,12 @@ describe AbstractJob do
 
     it 'returns @parser if @parser is set' do
       job.instance_variable_set(:@parser, 'a parser!')
-      job.parser.should eq 'a parser!'
+      expect(job.parser).to eq 'a parser!'
     end
 
     context 'with version_id' do
       it 'finds the parser by id' do
-        ParserVersion.should_receive(:find).with('666', params: { parser_id: '12345' })
+        expect(ParserVersion).to receive(:find).with('666', params: { parser_id: '12345' })
         job.parser
       end
     end
@@ -84,14 +85,14 @@ describe AbstractJob do
       end
 
       it 'finds the current parser version for the environment' do
-        ParserVersion.should_receive(:find).with(:one, from: :current, params: { parser_id: '12345', environment: 'staging' }) { version }
+        expect(ParserVersion).to receive(:find).with(:one, from: :current, params: { parser_id: '12345', environment: 'staging' }).and_return(version)
         job.parser
       end
 
       it 'should set the version_id of the fetched version' do
-        ParserVersion.should_receive(:find) { mock_model(ParserVersion, id: '888').as_null_object }
+        expect(ParserVersion).to receive(:find).and_return(mock_model(ParserVersion, id: '888').as_null_object)
         job.parser
-        job.version_id.should eq '888'
+        expect(job.version_id).to eq '888'
       end
 
       context 'the environment is preview and parser_code is present' do
@@ -99,16 +100,16 @@ describe AbstractJob do
         before(:each) do
           job.environment = 'preview'
           job.parser_code = 'new code'
-          Parser.stub(:find) { parser }
+          allow(Parser).to receive(:find).and_return(parser)
         end
 
         it 'finds the parser by id' do
-          Parser.should_receive(:find).with('12345')
+          expect(Parser).to receive(:find).with('12345')
           job.parser
         end
 
         it 'sets the parser content to parser_code' do
-          parser.should_receive(:content=).with('new code')
+          expect(parser).to receive(:content=).with('new code')
           job.parser
         end
       end
@@ -121,90 +122,90 @@ describe AbstractJob do
       end
 
       it 'finds the parser by id' do
-        Parser.should_receive(:find).with('12345')
+        expect(Parser).to receive(:find).with('12345')
         job.parser
       end
     end
   end
 
   describe '#required_enrichments' do
-    it 'should return an array of enrichments with required: true' do
-      job.stub_chain(:parser, :enrichment_definitions).and_return(ndha_rights: { required_for_active_record: true }, thumbnails: {})
-      job.required_enrichments.should eq [:ndha_rights]
+    it 'returns an array of enrichments with required: true' do
+      allow(job).to receive_message_chain(:parser, :enrichment_definitions).and_return(ndha_rights: { required_for_active_record: true }, thumbnails: {})
+      expect(job.required_enrichments).to eq [:ndha_rights]
     end
   end
 
   describe 'status' do
     describe 'initial' do
-      it 'should set the initial status to active' do
-        job.ready?.should be_truthy
+      it 'sets the initial status to active' do
+        expect(job.ready?).to be_truthy
       end
     end
 
     describe 'start!' do
-      it 'should set the start time to now' do
+      it 'sets the start time to now' do
         time = Time.now
         Timecop.freeze(time) do
           job.start!
-          job.start_time.to_i.should eq time.to_i
+          expect(job.start_time.to_i).to eq time.to_i
         end
       end
 
-      it 'should set the status to active' do
+      it 'sets the status to active' do
         job.start!
-        job.active?.should be_truthy
+        expect(job.active?).to be_truthy
       end
 
-      it 'should set the records count to 0' do
+      it 'sets the records count to 0' do
         job.start!
-        job.records_count.should eq 0
+        expect(job.records_count).to eq 0
       end
 
-      it 'should set the processed count to 0' do
+      it 'sets the processed count to 0' do
         job.start!
-        job.processed_count.should eq 0
+        expect(job.processed_count).to eq 0
       end
 
-      it 'should save the job' do
-        job.should_receive(:save)
+      it 'saves the job' do
+        expect(job).to receive(:save)
         job.start!
       end
     end
 
     describe 'finish!' do
-      it 'should set the status to finished' do
+      it 'sets the status to finished' do
         job.finish!
-        job.finished?.should be_truthy
+        expect(job.finished?).to be_truthy
       end
 
-      it 'should set the end time to now' do
+      it 'sets the end time to now' do
         time = Time.now
         Timecop.freeze(time) do
           job.finish!
-          job.end_time.to_i.should eq time.to_i
+          expect(job.end_time.to_i).to eq time.to_i
         end
       end
 
-      it 'should set the throughput' do
-        job.should_receive(:calculate_throughput)
+      it 'sets the throughput' do
+        expect(job).to receive(:calculate_throughput)
         job.finish!
       end
 
-      it 'should set the errors count' do
-        job.should_receive(:calculate_errors_count)
+      it 'sets the errors count' do
+        expect(job).to receive(:calculate_errors_count)
         job.finish!
       end
 
-      it 'should save the job' do
-        job.should_receive(:save)
+      it 'saves the job' do
+        expect(job).to receive(:save)
         job.finish!
       end
     end
 
     describe 'error!' do
-      it 'should set the status to failed' do
+      it 'sets the status to failed' do
         job.error!
-        job.failed?.should be_truthy
+        expect(job.failed?).to be_truthy
       end
 
       context 'start time' do
@@ -212,11 +213,11 @@ describe AbstractJob do
           time = Time.now
           Timecop.freeze(time) do
             job.error!
-            job.end_time.to_i.should eq time.to_i
+            expect(job.end_time.to_i).to eq time.to_i
           end
         end
 
-        it 'should not set the start time to now if set' do
+        it 'does not set the start time to now if set' do
           Timecop.freeze(Time.now - 1.hour) do
             job.start!
           end
@@ -224,27 +225,27 @@ describe AbstractJob do
           time = Time.now
           Timecop.freeze(time) do
             job.error!
-            job.reload.start_time.to_i.should_not eq time.to_i
+            expect(job.reload.start_time.to_i).to_not eq time.to_i
           end
         end
       end
 
-      it 'should set the end time to now' do
+      it 'sets the end time to now' do
         job.start!
         time = Time.now
         Timecop.freeze(time) do
           job.error!
-          job.end_time.to_i.should eq time.to_i
+          expect(job.end_time.to_i).to eq time.to_i
         end
       end
 
-      it 'should set the errors count' do
-        job.should_receive(:calculate_errors_count)
+      it 'sets the errors count' do
+        expect(job).to receive(:calculate_errors_count)
         job.error!
       end
 
-      it 'should save the job' do
-        job.should_receive(:save)
+      it 'saves the job' do
+        expect(job).to receive(:save)
         job.error!
       end
     end
@@ -254,21 +255,21 @@ describe AbstractJob do
         job.start!
       end
 
-      it 'should set the status to failed' do
+      it 'sets the status to failed' do
         job.stop!
-        job.stopped?.should be_truthy
+        expect(job.stopped?).to be_truthy
       end
 
-      it 'should set the end time to now' do
+      it 'sets the end time to now' do
         time = Time.now
         Timecop.freeze(time) do
           job.stop!
-          job.end_time.to_i.should eq time.to_i
+          expect(job.end_time.to_i).to eq time.to_i
         end
       end
 
-      it 'should save the job' do
-        job.should_receive(:save)
+      it 'saves the job' do
+        expect(job).to receive(:save)
         job.stop!
       end
     end
@@ -277,24 +278,24 @@ describe AbstractJob do
   describe 'test?' do
     it 'returns true' do
       job.environment = 'test'
-      job.test?.should be_truthy
+      expect(job.test?).to be_truthy
     end
 
     it 'returns false' do
       job.environment = 'staging'
-      job.test?.should be_falsey
+      expect(job.test?).to be_falsey
     end
   end
 
   describe '#preview?' do
     it 'returns true' do
       job.environment = 'preview'
-      job.preview?.should be_truthy
+      expect(job.preview?).to be_truthy
     end
 
     it 'returns false' do
       job.environment = 'staging'
-      job.preview?.should be_falsey
+      expect(job.preview?).to be_falsey
     end
   end
 
@@ -304,25 +305,25 @@ describe AbstractJob do
       job.status = 'finished'
     end
 
-    it 'should calculate the average record time' do
+    it 'calculates the average record time' do
       job.records_count = 100
-      job.stub(:duration) { 100 }
+      allow(job).to receive(:duration).and_return(100)
       job.calculate_throughput
-      job.throughput.should eq 1.0
+      expect(job.throughput).to eq 1.0
     end
 
     it 'returns 0 when records harvested is 0' do
       job.records_count = 0
-      job.stub(:duration) { 100 }
+      allow(job).to receive(:duration).and_return(100)
       job.calculate_throughput
-      job.throughput.should eq 0
+      expect(job.throughput).to eq 0
     end
 
     it 'should not return NaN' do
       job.records_count = 0
-      job.stub(:duration) { 0.0 }
+      allow(job).to receive(:duration).and_return(0.0)
       job.calculate_throughput
-      job.throughput.should be_nil
+      expect(job.throughput).to be_nil
     end
   end
 
@@ -332,29 +333,30 @@ describe AbstractJob do
     it 'returns a count of harvest jobs in the last 2 days' do
       old_finished_job = create(:abstract_job, status: 'finished', start_time: (DateTime.now - 3), environment: 'staging')
       since = DateTime.now - 2
-      AbstractJob.jobs_since('datetime' => since.to_s, 'environment' => 'staging', 'status' => 'finished').should eq [finished_job]
+      jobs_since = AbstractJob.jobs_since('datetime' => since.to_s, 'environment' => 'staging', 'status' => 'finished')
+      expect(jobs_since).to eq [finished_job]
     end
   end
 
   describe '#duration' do
     let!(:time) { Time.now }
 
-    it 'should return the duration in seconds' do
+    it 'returns the duration in seconds' do
       job.start_time = time - 10.seconds
       job.end_time = time
       job.save
       job.reload
-      job.duration.should eq 10.0
+      expect(job.duration).to eq 10.0
     end
 
     it 'returns nil start_time is nil' do
       job.start_time = nil
-      job.duration.should be_nil
+      expect(job.duration).to be_nil
     end
 
     it 'returns nil end_time is nil' do
       job.end_time = nil
-      job.duration.should be_nil
+      expect(job.duration).to be_nil
     end
 
     it 'returns the proper duration' do
@@ -362,66 +364,66 @@ describe AbstractJob do
       Timecop.freeze(time) do
         job = create(:abstract_job, start_time: time)
         job.end_time = Time.now + 5.seconds
-        job.duration.should eq 5
+        expect(job.duration).to eq 5
       end
     end
   end
 
   describe 'total_errors_count' do
-    it 'should return a sum of failed and invalid records' do
-      job.stub(:invalid_records) { double(:array, count: 10) }
-      job.stub(:failed_records) { double(:array, count: 20) }
-      job.total_errors_count.should eq 30
+    it 'returns a sum of failed and invalid records' do
+      allow(job).to receive(:invalid_records).and_return(double(:array, count: 10))
+      allow(job).to receive(:failed_records).and_return(double(:array, count: 20))
+      expect(job.total_errors_count).to eq 30
     end
   end
 
   describe '#errors_over_limit?' do
     context 'errors count over 100' do
-      before { job.stub(:total_errors_count) { 101 } }
+      before { allow(job).to receive(:total_errors_count).and_return(101) }
 
-      it 'should return true' do
-        job.errors_over_limit?.should be_truthy
+      it 'returns true' do
+        expect(job.errors_over_limit?).to be_truthy
       end
     end
 
     context 'errors count under 100' do
-      before { job.stub(:total_errors_count) { 99 } }
+      before { allow(job).to receive(:total_errors_count).and_return(99) }
 
-      it 'should return false' do
-        job.errors_over_limit?.should be_falsey
+      it 'returns false' do
+        expect(job.errors_over_limit?).to be_falsey
       end
     end
   end
 
-  describe 'clear_raw_data' do
-    it 'should remove invalid records' do
+  describe '#clear_raw_data' do
+    it 'removes invalid records' do
       job.invalid_records.create(raw_data: 'Wrong', errors_messages: [])
       job.clear_raw_data
       job.reload
-      job.invalid_records.count.should eq 0
+      expect(job.invalid_records.count).to eq 0
     end
 
-    it 'should remove failed records' do
+    it 'removes failed records' do
       job.failed_records.create(message: 'Hi')
       job.clear_raw_data
       job.reload
-      job.failed_records.count.should eq 0
+      expect(job.failed_records.count).to eq 0
     end
   end
 
   describe '#increment_records_count!' do
-    it 'should increment the records count' do
+    it 'increments the records count' do
       job.increment_records_count!
       job.reload
-      job.records_count.should eq 1
+      expect(job.records_count).to eq 1
     end
   end
 
   describe '#increment_processed_count!' do
-    it 'should increment the records count' do
+    it 'increments the records count' do
       job.increment_processed_count!
       job.reload
-      job.processed_count.should eq 1
+      expect(job.processed_count).to eq 1
     end
   end
 end
