@@ -3,45 +3,45 @@ require 'rails_helper'
 
 describe HarvestJob do
   context 'validations' do
-    it 'should not be possible to have 2 active jobs for the same parser/environment' do
+    it 'is not possible to have 2 active jobs for the same parser/environment' do
       job1 = create(:harvest_job, parser_id: '333', environment: 'staging', status: 'active')
       job2 = build(:harvest_job, parser_id: '333', environment: 'staging', status: 'active')
-      job2.should_not be_valid
+      expect(job2).to_not be_valid
     end
 
-    it 'should be possible to have 2 finished jobs for the same parser/environment' do
+    it 'is possible to have 2 finished jobs for the same parser/environment' do
       job1 = create(:harvest_job, parser_id: '333', environment: 'staging', status: 'finished')
       job2 = build(:harvest_job, parser_id: '333', environment: 'staging', status: 'finished')
-      job2.should be_valid
+      expect(job2).to be_valid
     end
 
-    it 'should be possible to enqueue a harvest and enrichment jobs simultaneously' do
+    it 'is possible to enqueue a harvest and enrichment jobs simultaneously' do
       job = create(:harvest_job, parser_id: '333', environment: 'staging', status: 'active')
       enrichment = build(:enrichment_job, parser_id: '333', environment: 'staging', status: 'active')
-      enrichment.should be_valid
+      expect(enrichment).to be_valid
     end
 
-    it "should not be possible to have a mode other then 'normal', 'full_and_flush' or 'incremental'" do
+    it "is not possible to have a mode other then 'normal', 'full_and_flush' or 'incremental'" do
       job1 = build(:harvest_job, parser_id: '333', mode: 'ya')
-      job1.should_not be_valid
+      expect(job1).to_not be_valid
     end
 
     %w[normal full_and_flush incremental].each do |mode|
-      it "should be valid with mode of '#{mode}'" do
+      it "is valid with mode of '#{mode}'" do
         job1 = build(:harvest_job, parser_id: '333', mode: mode)
-        job1.should be_valid
+        expect(job1).to be_valid
       end
     end
   end
 
   it 'enqueues a job after_create' do
-    HarvestWorker.should_receive(:perform_async)
+    expect(HarvestWorker).to receive(:perform_async)
     create(:harvest_job)
   end
 
   context 'preview environment' do
     it 'does not enque a job after create' do
-      HarvestWorker.should_not_receive(:perform_async)
+      expect(HarvestWorker).to_not receive(:perform_async)
       create(:harvest_job, environment: 'preview')
     end
   end
@@ -52,26 +52,26 @@ describe HarvestJob do
     let(:parser) { double(:parser, enrichment_definitions: { ndha_rights: proc {} }).as_null_object }
 
     before do
-      job.stub(:parser) { parser }
+      allow(job).to receive(:parser).and_return(parser)
     end
 
     it 'enqueues a enrichment job for each enrichment' do
       job.enrichments = ['ndha_rights']
-      EnrichmentJob.should_receive(:create_from_harvest_job).with(job, :ndha_rights)
+      expect(EnrichmentJob).to receive(:create_from_harvest_job).with(job, :ndha_rights)
       job.enqueue_enrichment_jobs
     end
 
     it 'should not enqueue enrichments not specified in the job' do
       job.enrichments = []
-      EnrichmentJob.should_not_receive(:create_from_harvest_job)
+      expect(EnrichmentJob).to_not receive(:create_from_harvest_job)
       job.enqueue_enrichment_jobs
     end
   end
 
   describe '#flush_old_records' do
-    it 'should post to the apis /harvester/records/flush action with source_id and the harvest_job_id' do
-      job.stub(:source_id) { 'source_id' }
-      RestClient.should_receive(:post).with("#{ENV['API_HOST']}/harvester/records/flush.json", source_id: 'source_id', job_id: job.id, api_key: ENV['HARVESTER_API_KEY'])
+    it 'post to the apis /harvester/records/flush action with source_id and the harvest_job_id' do
+      allow(job).to receive(:source_id).and_return('source_id')
+      expect(RestClient).to receive(:post).with("#{ENV['API_HOST']}/harvester/records/flush.json", source_id: 'source_id', job_id: job.id, api_key: ENV['HARVESTER_API_KEY'])
       job.flush_old_records
     end
   end
@@ -85,16 +85,16 @@ describe HarvestJob do
       SupplejackCommon.parser_base_path = Rails.root.to_s + '/tmp/parsers'
       parser.load_file('staging')
 
-      job.stub(:environment) { 'staging' }
-      job.stub(:parser) { parser }
-      job.stub(:finish!)
-      job.stub(:start!)
+      allow(job).to receive(:environment).and_return('staging')
+      allow(job).to receive(:parser).and_return(parser)
+      allow(job).to receive(:finish!)
+      allow(job).to receive(:start!)
 
-      LoadedParser::Staging::NatlibPages.stub(:environment=).with('staging')
-      LoadedParser::Staging::NatlibPages.stub(:records) { [record1, record2] }
+      allow(LoadedParser::Staging::NatlibPages).to receive(:environment=).with('staging')
+      allow(LoadedParser::Staging::NatlibPages).to receive(:records).and_return([record1, record2])
     end
 
-    it 'should start the job' do
+    it 'starts the job' do
       job.records { |r| r }
       expect(job).to have_received(:start!)
     end
@@ -105,9 +105,9 @@ describe HarvestJob do
     end
 
     it 'rescues exceptions from the whole harvest and stores it' do
-      LoadedParser::Staging::NatlibPages.stub(:records).and_raise 'Everything broke'
+      allow(LoadedParser::Staging::NatlibPages).to receive(:records).and_raise 'Everything broke'
       job.records { |r| r }
-      job.harvest_failure.message.should eq 'Everything broke'
+      expect(job.harvest_failure.message).to eq 'Everything broke'
     end
 
     it 'yields each record with index' do
@@ -122,31 +122,31 @@ describe HarvestJob do
 
     it 'flushes old records if full_and_flush is true' do
       job.records_count = 1
-      job.should_receive(:flush_old_records)
+      expect(job).to receive(:flush_old_records)
       job.finish!
     end
 
     it 'does not flush record if full_and_flush is false' do
       job.mode = 'normal'
-      job.should_not_receive(:flush_old_records)
+      expect(job).to_not receive(:flush_old_records)
       job.finish!
     end
 
     it 'does not flush record if limit is set' do
       job.limit = 100
-      job.should_not_receive(:flush_old_records)
+      expect(job).to_not receive(:flush_old_records)
       job.finish!
     end
 
     it 'does not flush records if a harvest failure occured' do
       job.build_harvest_failure
-      job.should_not_receive(:flush_old_records)
+      expect(job).to_not receive(:flush_old_records)
       job.finish!
     end
 
     it 'does not flush record if the job is manually stopped' do
       job.status = 'stopped'
-      job.should_not_receive(:flush_old_records)
+      expect(job).to_not receive(:flush_old_records)
       job.finish!
     end
   end
