@@ -7,39 +7,39 @@ describe LinkCheckJob do
   before { ENV['LINK_CHECKING_ENABLED'] = 'true' }
 
   describe 'validations' do
-    it 'should validates the presence of url' do
+    it 'validates the presence of url' do
       link_check_job.url = nil
-      link_check_job.should_not be_valid
+      expect(link_check_job).to_not be_valid
     end
 
-    it 'should validates the presence of record_id' do
+    it 'validatess the presence of record_id' do
       link_check_job.record_id = nil
-      link_check_job.should_not be_valid
+      expect(link_check_job).to_not be_valid
     end
 
-    it 'should validate the presence of source_id' do
+    it 'validates the presence of source_id' do
       link_check_job.source_id = nil
-      link_check_job.should_not be_valid
+      expect(link_check_job).to_not be_valid
     end
   end
 
   describe 'after:create' do
     after { link_check_job.save }
 
-    it 'should call enqueue' do
-      link_check_job.should_receive(:enqueue)
+    it 'calls enqueue' do
+      expect(link_check_job).to receive(:enqueue)
     end
   end
 
   describe '#enqueue' do
     it 'enqueues a job' do
-      LinkCheckWorker.should_receive(:perform_async).with(link_check_job.id.to_s)
+      expect(LinkCheckWorker).to receive(:perform_async).with(link_check_job.id.to_s)
       link_check_job.send(:enqueue)
     end
 
-    it 'should not enqueue a job if link checking is disabled' do
+    it 'does not enqueue a job if link checking is disabled' do
       ENV['LINK_CHECKING_ENABLED'] = nil
-      LinkCheckWorker.should_not_receive(:perform_async).with(link_check_job.id.to_s)
+      expect(LinkCheckWorker).to_not receive(:perform_async).with(link_check_job.id.to_s)
       link_check_job.send(:enqueue)
     end
   end
