@@ -12,27 +12,27 @@ describe ParserVersion do
     let!(:job1) { create(:harvest_job, start_time: time - 1.day, parser_id: '12', status: 'finished') }
     let!(:job2) { create(:harvest_job, start_time: time - 2.day, parser_id: '12', status: 'finished') }
 
-    it 'should return the last date a harvest job was run' do
+    it 'returns the last date a harvest job was run' do
       Timecop.freeze(time) do
         parser = Parser.new(id: '12', name: 'Europeana')
-        parser.last_harvested_at.to_i.should eq (time - 1.day).to_i
+        expect(parser.last_harvested_at.to_i).to eq (time - 1.day).to_i
       end
     end
 
-    it 'should return nil when no job has been run' do
+    it 'returns nil when no job has been run' do
       parser = Parser.new(id: '123', name: 'Europeana')
-      parser.last_harvested_at.should be_nil
+      expect(parser.last_harvested_at).to be_nil
     end
   end
 
   describe '#harvest_jobs' do
     it 'finds all the harvest jobs using this parser' do
-      HarvestJob.should_receive(:where).with(parser_id: parser_version.parser_id) { job }
+      expect(HarvestJob).to receive(:where).with(parser_id: parser_version.parser_id).and_return job
       parser_version.harvest_jobs
     end
 
     it 'finds all the harvest jobs with specified status' do
-      HarvestJob.should_receive(:where).with(parser_id: parser_version.parser_id, status: 'finished') { job }
+      expect(HarvestJob).to receive(:where).with(parser_id: parser_version.parser_id, status: 'finished') .and_return job
       parser_version.harvest_jobs('finished')
     end
   end
