@@ -7,12 +7,12 @@ describe Snippet do
   describe '.find_by_name' do
     it 'finds the snippet' do
       allow(Snippet).to receive(:find) { snippet }
-      Snippet.find_by_name('Copyright', :staging).should eq snippet
+      expect(Snippet.find_by_name('Copyright', :staging)).to eq snippet
     end
 
     it 'returns nil when a error is raised' do
-      Snippet.stub(:find).and_raise(ArgumentError)
-      Snippet.find_by_name('dsfsd', :staging).should be_nil
+      allow(Snippet).to receive(:find).and_raise(ArgumentError)
+      expect(Snippet.find_by_name('dsfsd', :staging)).to be_nil
     end
   end
 end
