@@ -5,26 +5,26 @@ describe AbstractWorker do
   let(:worker) { AbstractWorker.new }
   let(:job) { create(:harvest_job) }
 
-  before { AbstractJob.stub(:find) { job } }
+  before { allow(AbstractJob).to receive(:find) { job } }
 
   describe '#stop_harvest?' do
-    before { job.stub(:enqueue_enrichment_jobs) { nil } }
+    before { allow(job).to receive(:enqueue_enrichment_jobs) { nil } }
 
     context 'status is stopped' do
       let(:job) { create(:harvest_job, status: 'stopped') }
 
       it 'returns true' do
-        worker.stop_harvest?.should be_truthy
+        expect(worker.stop_harvest?).to be_truthy
       end
 
       it 'updates the job with the end time' do
-        job.should_receive(:finish!)
+        expect(job).to receive(:finish!)
         worker.stop_harvest?
       end
 
       it 'returns true true when errors over limit' do
-        job.stub(:errors_over_limit?) { true }
-        worker.stop_harvest?.should be_truthy
+        allow(job).to receive(:errors_over_limit?) { true }
+        expect(worker.stop_harvest?).to be_truthy
       end
     end
 
@@ -32,11 +32,11 @@ describe AbstractWorker do
       let(:job) { create(:harvest_job, status: 'finished') }
 
       it 'returns true' do
-        worker.stop_harvest?.should be_truthy
+        expect(worker.stop_harvest?).to be_truthy
       end
 
       it 'should not finsihed the job (again)' do
-        job.should_not_receive(:finish!)
+        expect(job).not_to receive(:finish!)
         worker.stop_harvest?
       end
     end
@@ -45,54 +45,54 @@ describe AbstractWorker do
       let(:job) { create(:harvest_job, status: 'active') }
 
       it 'returns true when errors over limit' do
-        job.stub(:errors_over_limit?) { true }
-        worker.stop_harvest?.should be_truthy
+        allow(job).to receive(:errors_over_limit?) { true }
+        expect(worker.stop_harvest?).to be_truthy
       end
 
       it 'returns false' do
-        worker.stop_harvest?.should be_falsey
+        expect(worker.stop_harvest?).to be_falsey
       end
     end
   end
 
   describe '#api_update_finished?' do
     it 'should return true if the api update is finished' do
-      job.stub(:posted_records_count) { 100 }
-      job.stub(:records_count) { 100 }
-      worker.send(:api_update_finished?).should be_truthy
+      allow(job).to receive(:posted_records_count) { 100 }
+      allow(job).to receive(:records_count) { 100 }
+      expect(worker.send(:api_update_finished?)).to be_truthy
     end
 
     it 'should return false if the api update is not finished' do
-      job.stub(:posted_records_count) { 10 }
-      job.stub(:records_count) { 100 }
-      worker.send(:api_update_finished?).should be_falsey
+      allow(job).to receive(:posted_records_count) { 10 }
+      allow(job).to receive(:records_count) { 100 }
+      expect(worker.send(:api_update_finished?)).to be_falsey
     end
 
     it 'should reload the enrichment job' do
-      job.should_receive(:reload)
+      expect(job).to receive(:reload)
       worker.send(:api_update_finished?)
     end
   end
 
   describe '#sanitize_id' do
     it 'accepts strings and returns the string' do
-      worker.send(:sanitize_id, 'abc').should eq 'abc'
+      expect(worker.send(:sanitize_id, 'abc')).to eq 'abc'
     end
 
     it 'it accepts serialized object_ids and returns the id string' do
-      worker.send(:sanitize_id, '$oid' => 'preview123').should eq 'preview123'
+      expect(worker.send(:sanitize_id, '$oid' => 'preview123')).to eq 'preview123'
     end
   end
 
   describe '#job' do
     it 'should find the job' do
       worker.instance_variable_set(:@job_id, 123)
-      AbstractJob.should_receive(:find).with('123') { job }
-      worker.job.should eq job
+      expect(AbstractJob).to receive(:find).with('123') { job }
+      expect(worker.job).to eq job
     end
 
     it 'memoizes the result' do
-      AbstractJob.should_receive(:find).once
+      expect(AbstractJob).to receive(:find).once
       worker.job
       worker.job
     end
